requirements.txt
"""
httpx==0.27.0
typer==0.12.3
pyyaml==6.0.1
"""
